Start addressing some comments
