+++ src/artemis.cpp
@@ -24,6 +24,7 @@
 #include "radiation/moments/moments.hpp"
 #include "radiation/radiation.hpp"
 #include "rotating_frame/rotating_frame.hpp"
+#include "self_gravity/self_gravity.hpp"
 #include "utils/artemis_utils.hpp"
 #include "utils/history.hpp"
 #include "utils/units.hpp"
@@ -90,6 +91,7 @@ Packages_t ProcessPackages(std::unique_ptr<ParameterInput> &pin) {
   const bool do_gas = pin->GetOrAddBoolean("physics", "gas", true);
   const bool do_dust = pin->GetOrAddBoolean("physics", "dust", false);
   const bool do_gravity = pin->GetOrAddBoolean("physics", "gravity", false);
+  const bool do_self_gravity = pin->GetOrAddBoolean("physics", "self_gravity", false);
   const bool do_nbody = pin->GetOrAddBoolean("physics", "nbody", false);
   const bool do_rotating_frame = pin->GetOrAddBoolean("physics", "rotating_frame", false);
   const bool do_cooling = pin->GetOrAddBoolean("physics", "cooling", false);
@@ -120,6 +122,7 @@ Packages_t ProcessPackages(std::unique_ptr<ParameterInput> &pin) {
   artemis->AddParam("do_gas", do_gas);
   artemis->AddParam("do_dust", do_dust);
   artemis->AddParam("do_gravity", do_gravity);
+  artemis->AddParam("do_self_gravity", do_self_gravity);
   artemis->AddParam("do_nbody", do_nbody);
   artemis->AddParam("do_rotating_frame", do_rotating_frame);
   artemis->AddParam("do_cooling", do_cooling);
@@ -145,6 +148,8 @@ Packages_t ProcessPackages(std::unique_ptr<ParameterInput> &pin) {
   // Call package initializers here
   if (do_nbody) packages.Add(NBody::Initialize(pin.get(), constants));
   if (do_gravity) packages.Add(Gravity::Initialize(pin.get(), constants, packages));
+  if (do_self_gravity)
+    packages.Add(SelfGravity::Initialize(pin.get(), constants, packages));
   if (do_gas) packages.Add(Gas::Initialize(pin.get(), units, constants, packages));
   if (do_dust) packages.Add(Dust::Initialize(pin.get(), units));
   if (do_rotating_frame) packages.Add(RotatingFrame::Initialize(pin.get()));
